Move some knowledge of session scope services provided by `core` back from `build-state` to `core`.
